test refactor to use base class
